% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmamba.R
\name{appmamba}
\alias{appmamba}
\alias{install_appmamba}
\alias{uninstall_appmamba}
\title{Manage Environment with \code{micromamba}}
\usage{
appmamba(...)

install_appmamba(force = FALSE)

uninstall_appmamba()
}
\arguments{
\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to
\code{micromamba}. Run \code{appmamba()} for more details.}

\item{force}{A logical value indicating whether to reinstall \code{appmamba}
if it is already installed.}
}
\description{
Manage Environment with \code{micromamba}
}
\section{Functions}{
\itemize{
\item \code{appmamba()}: \code{blit} utilizes \code{micromamba} to manage environments.
This function simply executes the specified \code{micromamba} command.

\item \code{install_appmamba()}: Install appmamba (\code{micromamba}).

\item \code{uninstall_appmamba()}: Remove appmamba (\code{micromamba}).

}}
\examples{
install_appmamba()
appmamba()
appmamba("env", "list")
uninstall_appmamba()
}
