% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-bedtools.R
\name{bedtools}
\alias{bedtools}
\title{Run bedtools}
\usage{
bedtools(subcmd = NULL, ..., bedtools = NULL)
}
\arguments{
\item{subcmd}{Sub-Command of bedtools. Details see: \code{cmd_help(bedtools())}.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{bedtools} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \code{cmd_help(bedtools())}.}

\item{bedtools}{A string of path to \code{bedtools} command.}
}
\value{
A \code{command} object.
}
\description{
The \code{bedtools} is a powerful toolset for genome arithmetic.
}
\seealso{
\itemize{
\item \url{http://bedtools.readthedocs.io/}
\item \url{https://github.com/arq5x/bedtools2/}
\item \code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}/\code{\link[=cmd_condaenv]{cmd_condaenv()}}
\item \code{\link[=cmd_on_start]{cmd_on_start()}}/\code{\link[=cmd_on_exit]{cmd_on_exit()}}
\item \code{\link[=cmd_on_succeed]{cmd_on_succeed()}}/\code{\link[=cmd_on_fail]{cmd_on_fail()}}
\item \code{\link[=cmd_parallel]{cmd_parallel()}}
}

Other \code{commands}:
\code{\link{allele_counter}()},
\code{\link{bcftools}()},
\code{\link{bowtie2}()},
\code{\link{bwa}()},
\code{\link{cellranger}()},
\code{\link{conda}()},
\code{\link{fastp}()},
\code{\link{fastq_pair}()},
\code{\link{gistic2}()},
\code{\link{kraken2}()},
\code{\link{kraken_tools}()},
\code{\link{perl}()},
\code{\link{pyscenic}()},
\code{\link{python}()},
\code{\link{samtools}()},
\code{\link{seqkit}()},
\code{\link{snpEff}()},
\code{\link{trust4}()},
\code{\link{varscan}()}
}
\concept{command}
