% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-bedtools_1.R
\name{bedtools_intersect}
\alias{bedtools_intersect}
\title{Run bedtools}
\usage{
bedtools_intersect(file1, file2, ofile, ..., `bedtools intersect` = NULL)
}
\arguments{
\item{file1}{Path to bed/gff(gtf)/vcf/bam(sam) file.}

\item{file2}{Path to the second file, some commands require.}

\item{ofile}{Path to the output file.}

\item{...}{\itemize{
\item \verb{bedtools intersect}: <\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \verb{bedtools intersect} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \verb{cmd_help(bedtools intersect())}.
\item \verb{bedtools merge}: <\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \verb{bedtools merge} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \verb{cmd_help(bedtools merge())}.
\item \verb{bedtools coverage}: <\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \verb{bedtools coverage} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \verb{cmd_help(bedtools coverage())}.
\item \verb{bedtools getfasta}: <\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \verb{bedtools getfasta} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \verb{cmd_help(bedtools getfasta())}.
\item \verb{bedtools closest}: <\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \verb{bedtools closest} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \verb{cmd_help(bedtools closest())}.
}}

\item{-}{\verb{bedtools intersect}: A string of path to \verb{bedtools intersect} command.
\itemize{
\item \verb{bedtools merge}: A string of path to \verb{bedtools merge} command.
\item \verb{bedtools coverage}: A string of path to \verb{bedtools coverage} command.
\item \verb{bedtools getfasta}: A string of path to \verb{bedtools getfasta} command.
\item \verb{bedtools closest}: A string of path to \verb{bedtools closest} command.
}}
}
\value{
A \code{command} object.
}
\description{
The \code{bedtools} is a powerful toolset for genome arithmetic,
which includes \code{intersect}, \code{merge}, \code{coverage}, \code{getfasta}, \code{closest}, etc.
}
\seealso{
\itemize{
\item \url{https://github.com/arq5x/bedtools2/}
\item \code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}/\code{\link[=cmd_condaenv]{cmd_condaenv()}}
\item \code{\link[=cmd_on_start]{cmd_on_start()}}/\code{\link[=cmd_on_exit]{cmd_on_exit()}}
\item \code{\link[=cmd_on_succeed]{cmd_on_succeed()}}/\code{\link[=cmd_on_fail]{cmd_on_fail()}}
\item \code{\link[=cmd_parallel]{cmd_parallel()}}
}

Other \code{commands}:
\code{\link{allele_counter}()},
\code{\link{bedtools}()},
\code{\link{cellranger}()},
\code{\link{conda}()},
\code{\link{fastp}()},
\code{\link{fastq_pair}()},
\code{\link{gistic2}()},
\code{\link{kraken2}()},
\code{\link{kraken_tools}()},
\code{\link{perl}()},
\code{\link{pyscenic}()},
\code{\link{python}()},
\code{\link{samtools}()},
\code{\link{seqkit}()},
\code{\link{trust4}()}
}
\concept{command}
