% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blit-package.R
\docType{package}
\name{blit-package}
\alias{blit}
\alias{blit-package}
\title{blit: Bioinformatics Library for Integrated Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An all-encompassing R toolkit designed to streamline the process of calling various bioinformatics software and then performing data analysis and visualization in R. With 'blit', users can easily integrate a wide array of bioinformatics command line tools into their workflows, leveraging the power of R for sophisticated data manipulation and graphical representation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WangLabCSU/blit}
  \item Report bugs at \url{https://github.com/WangLabCSU/blit/issues}
}

}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunp96@163.com} (\href{https://orcid.org/0000-0003-2801-3332}{ORCID})

Authors:
\itemize{
  \item Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})
}

Other contributors:
\itemize{
  \item Jennifer Lu \email{jennifer.lu717@gmail.com} (Author of the included scripts from Kraken2 and KrakenTools libraries) [copyright holder]
  \item Li Song \email{Li.Song@dartmouth.edu} (Author of included scripts from TRUST4 library) [copyright holder]
  \item X. Shirley Liu \email{xsliu@ds.dfci.harvard.edu} (Author of included scripts from TRUST4 library) [copyright holder]
}

}
\keyword{internal}
