% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-bowtie2.R
\name{bowtie2}
\alias{bowtie2}
\title{Run bowtie2}
\usage{
bowtie2(index, reads, ofile, ..., bowtie2 = NULL)
}
\arguments{
\item{index}{Path to bowtie2 index prefix (without file extensions).}

\item{reads}{A character vector of FASTQ files used as input to bowtie2.}

\item{ofile}{A string of path to the output sam file.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{bowtie2} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \code{cmd_help(bowtie2())}.}

\item{bowtie2}{A string of path to \code{bowtie2} command.}
}
\value{
A \code{command} object.
}
\description{
Bowtie 2 is an ultrafast and memory-efficient tool
for aligning sequencing reads to long reference sequences.
It is particularly good at aligning reads of about 50 up to 100s or 1,000s of characters,
and particularly good at aligning to relatively long (e.g. mammalian) genomes.
Bowtie 2 indexes the genome with an FM Index to keep its memory footprint small:
for the human genome, its memory footprint is typically around 3.2 GB.
Bowtie 2 supports gapped, local, and paired-end alignment modes.
}
\seealso{
\itemize{
\item \url{https://bowtie-bio.sourceforge.net/bowtie2/index.shtml}
\item \code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}/\code{\link[=cmd_condaenv]{cmd_condaenv()}}
\item \code{\link[=cmd_on_start]{cmd_on_start()}}/\code{\link[=cmd_on_exit]{cmd_on_exit()}}
\item \code{\link[=cmd_on_succeed]{cmd_on_succeed()}}/\code{\link[=cmd_on_fail]{cmd_on_fail()}}
\item \code{\link[=cmd_parallel]{cmd_parallel()}}
}

Other \code{commands}:
\code{\link{allele_counter}()},
\code{\link{bcftools}()},
\code{\link{cellranger}()},
\code{\link{conda}()},
\code{\link{fastp}()},
\code{\link{fastq_pair}()},
\code{\link{gistic2}()},
\code{\link{kraken2}()},
\code{\link{kraken_tools}()},
\code{\link{perl}()},
\code{\link{pyscenic}()},
\code{\link{python}()},
\code{\link{samtools}()},
\code{\link{seqkit}()},
\code{\link{snpEff}()},
\code{\link{trust4}()},
\code{\link{varscan}()}
}
\concept{command}
