% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-bwa.R
\name{bwa}
\alias{bwa}
\title{Run BWA}
\usage{
bwa(subcmd = NULL, ..., bwa = NULL)
}
\arguments{
\item{subcmd}{Sub-Command of BWA (e.g., "index", "mem").}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{bwa} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \code{cmd_help(bwa())}.}

\item{bwa}{A string of path to \code{bwa} command.}
}
\value{
A \code{command} object.
}
\description{
BWA is a software package that aligns low-divergence sequences to a
large reference genome, such as the human genome
}
\examples{
\dontrun{
# Index reference genome
bwa("index", "-a", "bwtsw", "reference.fa") |>
  cmd_run()

# Paired-end sequence alignment
bwa("mem", "-t", "4", "reference.fa", "read1.fq", "read2.fq") |>
  cmd_run(stdout = "output.sam")

# Single-end alignment (generate sai file)
bwa("aln", "-t", "4", "reference.fa", "read.fq") |>
  cmd_run(stdout = "read.sai")
}
}
\seealso{
\itemize{
\item \url{http://bio-bwa.sourceforge.net/}
\item \code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}/\code{\link[=cmd_condaenv]{cmd_condaenv()}}
\item \code{\link[=cmd_on_start]{cmd_on_start()}}/\code{\link[=cmd_on_exit]{cmd_on_exit()}}
\item \code{\link[=cmd_on_succeed]{cmd_on_succeed()}}/\code{\link[=cmd_on_fail]{cmd_on_fail()}}
\item \code{\link[=cmd_parallel]{cmd_parallel()}}
}

Other \code{commands}:
\code{\link{allele_counter}()},
\code{\link{bcftools}()},
\code{\link{bedtools}()},
\code{\link{bowtie2}()},
\code{\link{cellranger}()},
\code{\link{conda}()},
\code{\link{fastp}()},
\code{\link{fastq_pair}()},
\code{\link{gistic2}()},
\code{\link{kraken2}()},
\code{\link{kraken_tools}()},
\code{\link{perl}()},
\code{\link{pyscenic}()},
\code{\link{python}()},
\code{\link{samtools}()},
\code{\link{seqkit}()},
\code{\link{snpEff}()},
\code{\link{trust4}()},
\code{\link{varscan}()}
}
\concept{command}
