% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-setup.R
\name{cmd_conda}
\alias{cmd_conda}
\title{Set \code{conda-like} environment prefix to the \code{PATH} environment variables}
\usage{
cmd_conda(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=cmd_condaenv]{cmd_condaenv()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Set \code{conda-like} environment prefix to the
\code{PATH} environment variables
}
