% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-fastp.R
\name{fastp}
\alias{fastp}
\title{Run fastp}
\usage{
fastp(fq1, ofile1, ..., fq2 = NULL, ofile2 = NULL, fastp = NULL)
}
\arguments{
\item{fq1, fq2}{A string of fastq file path.}

\item{ofile1, ofile2}{A string of path to the output fastq file.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{fastp} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \code{cmd_help(fastp())}.}

\item{fastp}{A string of path to \code{fastp} command.}
}
\value{
A \code{command} object.
}
\description{
The \code{fastp} is a tool designed to provide ultrafast all-in-one preprocessing
and quality control for FastQ data.
}
\seealso{
\itemize{
\item \url{https://github.com/OpenGene/fastp}
\item \code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}/\code{\link[=cmd_condaenv]{cmd_condaenv()}}
\item \code{\link[=cmd_on_start]{cmd_on_start()}}/\code{\link[=cmd_on_exit]{cmd_on_exit()}}
\item \code{\link[=cmd_on_succeed]{cmd_on_succeed()}}/\code{\link[=cmd_on_fail]{cmd_on_fail()}}
\item \code{\link[=cmd_parallel]{cmd_parallel()}}
}

Other \code{commands}:
\code{\link{allele_counter}()},
\code{\link{bcftools}()},
\code{\link{cellranger}()},
\code{\link{conda}()},
\code{\link{fastq_pair}()},
\code{\link{gistic2}()},
\code{\link{kraken2}()},
\code{\link{kraken_tools}()},
\code{\link{perl}()},
\code{\link{pyscenic}()},
\code{\link{python}()},
\code{\link{samtools}()},
\code{\link{seqkit}()},
\code{\link{snpEff}()},
\code{\link{trust4}()},
\code{\link{varscan}()}
}
\concept{command}
