% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-freebayes.R
\name{freebayes}
\alias{freebayes}
\title{Run freebayes}
\usage{
freebayes(ref, input, ofile, ..., freebayes = NULL)
}
\arguments{
\item{ref}{A string of reference file path.}

\item{input}{A string of path to the input bam file.}

\item{ofile}{A string of path to the output vcf file.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{freebayes} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \code{cmd_help(freebayes())}.}

\item{freebayes}{A string of path to \code{freebayes} command.}
}
\value{
A \code{command} object.
}
\description{
freebayes is a Bayesian genetic variant detector designed to find small polymorphisms,
specifically SNPs (single-nucleotide polymorphisms), indels (insertions and deletions),
MNPs (multi-nucleotide polymorphisms), and complex events (composite insertion and substitution events)
smaller than the length of a short-read sequencing alignment.
}
\seealso{
\itemize{
\item \url{https://github.com/freebayes/freebayes}
\item \code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}/\code{\link[=cmd_condaenv]{cmd_condaenv()}}
\item \code{\link[=cmd_on_start]{cmd_on_start()}}/\code{\link[=cmd_on_exit]{cmd_on_exit()}}
\item \code{\link[=cmd_on_succeed]{cmd_on_succeed()}}/\code{\link[=cmd_on_fail]{cmd_on_fail()}}
\item \code{\link[=cmd_parallel]{cmd_parallel()}}
}

Other \code{commands}:
\code{\link{allele_counter}()},
\code{\link{bcftools}()},
\code{\link{bedtools}()},
\code{\link{bowtie2}()},
\code{\link{bwa}()},
\code{\link{cellranger}()},
\code{\link{conda}()},
\code{\link{fastp}()},
\code{\link{fastq_pair}()},
\code{\link{gistic2}()},
\code{\link{kraken2}()},
\code{\link{kraken_tools}()},
\code{\link{minimap2}()},
\code{\link{perl}()},
\code{\link{pyscenic}()},
\code{\link{python}()},
\code{\link{samtools}()},
\code{\link{seqkit}()},
\code{\link{snpEff}()},
\code{\link{strelka}()},
\code{\link{trust4}()},
\code{\link{varscan}()}
}
\concept{command}
