% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-.R
\name{make_command}
\alias{make_command}
\title{Helper function to create new command.}
\usage{
make_command(name, fun, envir = caller_env())
}
\arguments{
\item{name}{A string of the function name.}

\item{fun}{A function used to initialize the \code{\link{Command}} object.}

\item{envir}{A environment used to bind the created function.}
}
\value{
A function.
}
\description{
\code{make_command} is a helper function used by developers to create function for
a new \code{\link{Command}} object. It should not be used by end users.
}
\seealso{
\code{\link{Command}}
}
