% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-minimap2.R
\name{minimap2}
\alias{minimap2}
\title{Run minimap2}
\usage{
minimap2(method, ref, reads, ofile, ..., minimap2 = NULL)
}
\arguments{
\item{method}{Mapping preset: "map-ont" (single-end long reads) or "sr" (paired-end short reads).}

\item{ref}{Path to the reference genome FASTA file.}

\item{reads}{A character vector of FASTQ files used as input to minimap2.}

\item{ofile}{A string of path to the output SAM file.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{minimap2} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \code{cmd_help(minimap2())}.}

\item{minimap2}{A string of path to \code{minimap2} command.}
}
\value{
A \code{command} object.
}
\description{
Minimap2 is a versatile sequence alignment program that aligns DNA or mRNA sequences against a large reference database.
Typical use cases include:
(1) mapping PacBio or Oxford Nanopore genomic reads to the human genome;
(2) finding overlaps between long reads with error rate up to ~15\%;
(3) splice-aware alignment of PacBio Iso-Seq or Nanopore cDNA or Direct RNA reads against a reference genome;
(4) aligning Illumina single- or paired-end reads;
(5) assembly-to-assembly alignment;
(6) full-genome alignment between two closely related species with divergence below ~15\%.
}
\seealso{
\itemize{
\item \url{https://github.com/lh3/minimap2}
\item \code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}/\code{\link[=cmd_condaenv]{cmd_condaenv()}}
\item \code{\link[=cmd_on_start]{cmd_on_start()}}/\code{\link[=cmd_on_exit]{cmd_on_exit()}}
\item \code{\link[=cmd_on_succeed]{cmd_on_succeed()}}/\code{\link[=cmd_on_fail]{cmd_on_fail()}}
\item \code{\link[=cmd_parallel]{cmd_parallel()}}
}

Other \code{commands}:
\code{\link{allele_counter}()},
\code{\link{bcftools}()},
\code{\link{bedtools}()},
\code{\link{bowtie2}()},
\code{\link{bwa}()},
\code{\link{cellranger}()},
\code{\link{conda}()},
\code{\link{fastp}()},
\code{\link{fastq_pair}()},
\code{\link{freebayes}()},
\code{\link{gistic2}()},
\code{\link{kraken2}()},
\code{\link{kraken_tools}()},
\code{\link{perl}()},
\code{\link{pyscenic}()},
\code{\link{python}()},
\code{\link{samtools}()},
\code{\link{seqkit}()},
\code{\link{snpEff}()},
\code{\link{strelka}()},
\code{\link{trust4}()},
\code{\link{varscan}()}
}
\concept{command}
