% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-perl.R
\name{perl}
\alias{perl}
\title{Perl is a highly capable, feature-rich programming language with over 36
years of development.}
\usage{
perl(..., perl = NULL)
}
\arguments{
\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{perl} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \code{cmd_help(perl())}.}

\item{perl}{A string of path to \code{perl} command.}
}
\value{
A \code{command} object.
}
\description{
Perl is a highly capable, feature-rich programming language with over 36
years of development.
}
\seealso{
\itemize{
\item \url{https://www.perl.org/}
\item \code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}/\code{\link[=cmd_conda]{cmd_conda()}}
\item \code{\link[=cmd_on_start]{cmd_on_start()}}/\code{\link[=cmd_on_exit]{cmd_on_exit()}}
\item \code{\link[=cmd_on_succeed]{cmd_on_succeed()}}/\code{\link[=cmd_on_fail]{cmd_on_fail()}}
\item \code{\link[=cmd_parallel]{cmd_parallel()}}
}

Other \code{commands}:
\code{\link{allele_counter}()},
\code{\link{cellranger}()},
\code{\link{conda}()},
\code{\link{fastq_pair}()},
\code{\link{gistic2}()},
\code{\link{kraken2}()},
\code{\link{kraken_tools}()},
\code{\link{pyscenic}()},
\code{\link{python}()},
\code{\link{samtools}()},
\code{\link{seqkit}()},
\code{\link{trust4}()}
}
\concept{command}
