% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-snpEff.R
\name{snpEff}
\alias{snpEff}
\title{Genetic variant annotation, and functional effect prediction toolbox.
It annotates and predicts the effects of genetic variants on genes and proteins
(such as amino acid changes).}
\usage{
snpEff(subcmd = NULL, ..., snpEff = NULL)
}
\arguments{
\item{subcmd}{Sub-Command of snpEff. Details see: \code{cmd_help(snpEff())}.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{snpEff} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \code{cmd_help(snpEff())}.}

\item{snpEff}{A string of path to \code{snpEff} command.}
}
\value{
A \code{command} object.
}
\description{
Genetic variant annotation, and functional effect prediction toolbox.
It annotates and predicts the effects of genetic variants on genes and proteins
(such as amino acid changes).
}
\seealso{
\itemize{
\item \url{https://pcingola.github.io/SnpEff/}
\item \code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}/\code{\link[=cmd_condaenv]{cmd_condaenv()}}
\item \code{\link[=cmd_on_start]{cmd_on_start()}}/\code{\link[=cmd_on_exit]{cmd_on_exit()}}
\item \code{\link[=cmd_on_succeed]{cmd_on_succeed()}}/\code{\link[=cmd_on_fail]{cmd_on_fail()}}
\item \code{\link[=cmd_parallel]{cmd_parallel()}}
}

Other \code{commands}:
\code{\link{allele_counter}()},
\code{\link{cellranger}()},
\code{\link{conda}()},
\code{\link{fastp}()},
\code{\link{fastq_pair}()},
\code{\link{gistic2}()},
\code{\link{kraken2}()},
\code{\link{kraken_tools}()},
\code{\link{perl}()},
\code{\link{pyscenic}()},
\code{\link{python}()},
\code{\link{samtools}()},
\code{\link{seqkit}()},
\code{\link{trust4}()}
}
\concept{command}
