% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-strelka.R
\name{strelka}
\alias{strelka}
\title{Run strelka}
\usage{
strelka(script, ...)
}
\arguments{
\item{script}{Name of the Strelka script. One of
\code{"configureStrelkaGermlineWorkflow"}, \code{"configureStrelkaSomaticWorkflow"}, and \code{"runWorkflow"}.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{strelka} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \code{cmd_help(strelka())}.}

\item{strelka}{A string of path to \code{strelka} command.}
}
\value{
A \code{command} object.
}
\description{
Strelka is a fast and accurate small variant caller optimized for analysis of germline variation
in small cohorts and somatic variation in tumor/normal sample pairs.
}
\seealso{
\itemize{
\item \url{https://github.com/Illumina/strelka}
\item \code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}/\code{\link[=cmd_condaenv]{cmd_condaenv()}}
\item \code{\link[=cmd_on_start]{cmd_on_start()}}/\code{\link[=cmd_on_exit]{cmd_on_exit()}}
\item \code{\link[=cmd_on_succeed]{cmd_on_succeed()}}/\code{\link[=cmd_on_fail]{cmd_on_fail()}}
\item \code{\link[=cmd_parallel]{cmd_parallel()}}
}

Other \code{commands}:
\code{\link{allele_counter}()},
\code{\link{bcftools}()},
\code{\link{bedtools}()},
\code{\link{bowtie2}()},
\code{\link{bwa}()},
\code{\link{cellranger}()},
\code{\link{conda}()},
\code{\link{fastp}()},
\code{\link{fastq_pair}()},
\code{\link{freebayes}()},
\code{\link{gistic2}()},
\code{\link{kraken2}()},
\code{\link{kraken_tools}()},
\code{\link{minimap2}()},
\code{\link{perl}()},
\code{\link{pyscenic}()},
\code{\link{python}()},
\code{\link{samtools}()},
\code{\link{seqkit}()},
\code{\link{snpEff}()},
\code{\link{trust4}()},
\code{\link{varscan}()}
}
\concept{command}
