% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-varscan.R
\name{varscan}
\alias{varscan}
\title{VarScan is a platform-independent software tool
developed at the Genome Institute at Washington University to detect variants in NGS data.}
\usage{
varscan(subcmd = NULL, ..., varscan = NULL)
}
\arguments{
\item{subcmd}{Sub-Command of varscan. Details see: \code{cmd_help(varscan())}.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{varscan} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \code{cmd_help(varscan())}.}

\item{varscan}{A string of path to \code{varscan} command.}
}
\value{
A \code{command} object.
}
\description{
VarScan is a platform-independent software tool
developed at the Genome Institute at Washington University to detect variants in NGS data.
}
\seealso{
\itemize{
\item \url{https://varscan.sourceforge.net/}
\item \code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}/\code{\link[=cmd_condaenv]{cmd_condaenv()}}
\item \code{\link[=cmd_on_start]{cmd_on_start()}}/\code{\link[=cmd_on_exit]{cmd_on_exit()}}
\item \code{\link[=cmd_on_succeed]{cmd_on_succeed()}}/\code{\link[=cmd_on_fail]{cmd_on_fail()}}
\item \code{\link[=cmd_parallel]{cmd_parallel()}}
}

Other \code{commands}:
\code{\link{allele_counter}()},
\code{\link{bcftools}()},
\code{\link{bedtools}()},
\code{\link{bowtie2}()},
\code{\link{cellranger}()},
\code{\link{conda}()},
\code{\link{fastp}()},
\code{\link{fastq_pair}()},
\code{\link{gistic2}()},
\code{\link{kraken2}()},
\code{\link{kraken_tools}()},
\code{\link{perl}()},
\code{\link{pyscenic}()},
\code{\link{python}()},
\code{\link{samtools}()},
\code{\link{seqkit}()},
\code{\link{snpEff}()},
\code{\link{trust4}()}
}
\concept{command}
