% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-cellranger.R
\name{cellranger}
\alias{cellranger}
\title{Run cellranger}
\usage{
cellranger(subcmd = NULL, ..., cellranger = NULL)
}
\arguments{
\item{subcmd}{Sub-Command of cellranger.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \verb{cellranger subcmd} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link{shQuote}}. Details see: \verb{cmd_help(cellranger subcmd())}.}

\item{cellranger}{A string of path to \code{cellranger} command.}
}
\value{
A \code{command} object.
}
\description{
Run cellranger
}
\seealso{
Other \code{commands}:
\code{\link{allele_counter}()},
\code{\link{fastq_pair}()},
\code{\link{gistic2}()},
\code{\link{kraken2}()},
\code{\link{kraken_tools}()},
\code{\link{perl}()},
\code{\link{pyscenic}()},
\code{\link{python}()},
\code{\link{seqkit}()},
\code{\link{trust4}()}
}
\concept{command}
