% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-cellranger.R
\name{cellranger}
\alias{cellranger}
\title{Run cellranger}
\usage{
cellranger(subcmd = NULL, ..., cellranger = NULL)
}
\arguments{
\item{subcmd}{Sub-Command of cellranger.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \verb{cellranger subcmd} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link{shQuote}}. Details see: \verb{cmd_help(cellranger subcmd())}.}

\item{cellranger}{A string of path to \code{cellranger} command.}
}
\value{
A \code{command} object.
}
\description{
Run cellranger
}
\seealso{
\itemize{
\item \code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}
\item \code{\link[=cmd_run]{cmd_run()}}/\code{\link[=cmd_background]{cmd_background()}}/\code{\link[=cmd_help]{cmd_help()}}
}
}
