% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-cellranger.R
\name{cellranger}
\alias{cellranger}
\title{Run cellranger}
\usage{
cellranger(subcmd = NULL, ..., cellranger = NULL)
}
\arguments{
\item{subcmd}{Sub-Command of cellranger.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{cellranger} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \code{cmd_help(cellranger())}.}

\item{cellranger}{A string of path to \code{cellranger} command.}
}
\value{
A \code{command} object.
}
\description{
Run cellranger
}
\examples{
\dontrun{
fastq_dir  # 10x raw fastq files directory
genome_ref # Please download the transcriptome reference data
cellranger(
    "count",
    sprintf("--fastqs=\%s", fastq_dir),
    sprintf("--id=\%s", basename(fastq_dir)),
    sprintf("--sample=\%s", basename(fastq_dir)),
    sprintf("--localcores=\%s", parallel::detectCores()),
    sprintf("--transcriptome=\%s", genome_ref),
    sprintf("--chemistry=\%s", shQuote("auto")),
    "--nosecondary"
)
}
}
\seealso{
\itemize{
\item \url{https://github.com/10XGenomics/cellranger}
\item \url{https://www.10xgenomics.com/support/software/cell-ranger/latest}
\item \url{https://www.10xgenomics.com/support/software/cell-ranger/downloads#reference-downloads}
\item \code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}/\code{\link[=cmd_conda]{cmd_conda()}}
\item \code{\link[=cmd_on_start]{cmd_on_start()}}/\code{\link[=cmd_on_exit]{cmd_on_exit()}}
\item \code{\link[=cmd_on_succeed]{cmd_on_succeed()}}/\code{\link[=cmd_on_fail]{cmd_on_fail()}}
\item \code{\link[=cmd_parallel]{cmd_parallel()}}
}

Other \code{commands}:
\code{\link{allele_counter}()},
\code{\link{conda}()},
\code{\link{fastq_pair}()},
\code{\link{gistic2}()},
\code{\link{kraken2}()},
\code{\link{kraken_tools}()},
\code{\link{perl}()},
\code{\link{pyscenic}()},
\code{\link{python}()},
\code{\link{samtools}()},
\code{\link{seqkit}()},
\code{\link{trust4}()}
}
\concept{command}
