% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-run.R
\name{cmd_run}
\alias{cmd_run}
\alias{cmd_help}
\alias{cmd_background}
\title{Execute command}
\usage{
cmd_run(
  command,
  stdout = TRUE,
  stderr = TRUE,
  stdin = NULL,
  stdout_callback = NULL,
  stderr_callback = NULL,
  timeout = NULL,
  spinner = FALSE,
  verbose = TRUE
)

cmd_help(
  command,
  stdout,
  stderr,
  stdout_callback = NULL,
  stderr_callback = NULL,
  verbose = TRUE
)

cmd_background(command, stdout, stderr, stdin = NULL, verbose = TRUE)
}
\arguments{
\item{command}{A \code{command} object.}

\item{stdout, stderr}{How output streams of the child process are processed.
Possible values are:
\itemize{
\item \code{TRUE}: Print the child output in R console.
\item \code{FALSE}: Suppress output stream
\item \strong{string}: Name or path of file to redirect output. Note that if you
specify a relative path, it will be relative to the current working
directory, even if you specify another directory using \code{\link[=cmd_wd]{cmd_wd()}}.
\item \code{connection}: A writable R \code{\link{connection}} object.
}

For \code{cmd_help()}, only a string (file path), or \code{connection} can be used.

For \code{cmd_background()}, only a string (file path), or a single boolean value
can be used.

\code{stderr} can also accept \code{NULL}, which means redirect it to the same
connection (i.e. pipe or file) as \code{stdout}.}

\item{stdin}{should the input be diverted? A character string naming a file.}

\item{stdout_callback, stderr_callback}{\code{NULL}, or a function to call for
every line of the standard output/error.}

\item{timeout}{Timeout in seconds. This is a limit for the elapsed time
running command in the separate process.}

\item{spinner}{Whether to show a reassuring spinner while the process
is running.}

\item{verbose}{A single boolean value indicating whether the command
execution should be verbose.}
}
\value{
\itemize{
\item \code{cmd_run}: Exit status invisiblely.
}

\itemize{
\item \code{cmd_help}: The input \code{command} invisiblely.
}

\itemize{
\item \code{cmd_background}: A \code{\link[processx:process]{process}} object.
}
}
\description{
\itemize{
\item \code{cmd_run}: Run the command.
\item \code{cmd_help}: Print the help document for this command.
\item \code{cmd_background}: Run the command in the background.
}
}
\seealso{
\code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}
}
