% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-.R
\name{cmd_run}
\alias{cmd_run}
\alias{cmd_background}
\alias{cmd_help}
\title{Execute command}
\usage{
cmd_run(
  command,
  stdout = TRUE,
  stderr = TRUE,
  stdin = NULL,
  timeout = NULL,
  verbose = TRUE
)

cmd_background(
  command,
  stdout = NULL,
  stderr = NULL,
  stdin = NULL,
  verbose = TRUE
)

cmd_help(command, stdout = NULL, stderr = NULL, verbose = TRUE)
}
\arguments{
\item{command}{A \code{command} object.}

\item{stdout, stderr}{How output streams of the child process are processed.
Possible values are:
\itemize{
\item \code{TRUE}: Print the child output in R console.
\item \code{FALSE}: Suppress output stream
\item \strong{string}: name or path of file to redirect output
}

For \code{cmd_background()} and \code{cmd_help()}, only a string (file path) can be
used.}

\item{stdin}{should the input be diverted? A character string naming a file.}

\item{timeout}{Timeout in seconds. This is a limit for the elapsed time
running command in the separate process.}

\item{verbose}{A single boolean value indicates whether to print running
command message.}
}
\value{
\itemize{
\item \code{cmd_run}: Exit status.
}

\itemize{
\item \code{cmd_background}: The process id, the process can be killed manually with
\code{\link[tools:pskill]{tools::pskill}}.
}

\itemize{
\item \code{cmd_help}: the input \code{command} invisiblely.
}
}
\description{
\itemize{
\item \code{cmd_run}: Run the command.
\item \code{cmd_background}: Run the command in the background using
\code{parallel::mcparallel}, not available on windows.
\item \code{cmd_help}: Print the help document for this command.
}
}
\seealso{
\code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}
}
