% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-context.R
\name{cmd_wd}
\alias{cmd_wd}
\alias{cmd_envvar}
\alias{cmd_envpath}
\alias{cmd_on_start}
\alias{cmd_on_exit}
\alias{cmd_on_fail}
\title{Setup the context for the command}
\usage{
cmd_wd(command, wd = NULL)

cmd_envvar(command, ..., action = "replace", sep = NULL)

cmd_envpath(command, ..., action = "prefix", name = "PATH")

cmd_on_start(command, ...)

cmd_on_exit(command, ...)

cmd_on_fail(command, ...)
}
\arguments{
\item{command}{A \code{command} object.}

\item{wd}{A string or \code{NULL} define the working directory of the command.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}>:
\itemize{
\item \code{cmd_envvar}: Named character define the environment variables.
\item \code{cmd_envpath}: Unnamed character to define the \code{PATH}-like environment
variables \code{name}.
\item \code{cmd_on_start}: Expression to be evaluated when the command started.
\item \code{cmd_on_exit}: Expression to be evaluated when the command finished.
\item \code{cmd_on_fail}: Expression to be evaluated when the command failed.
}}

\item{action}{Should the new values \code{"replace"}, \code{"prefix"} or \code{"suffix"}
existing environment variables?}

\item{sep}{A string to separate new and old value when \code{action} is \code{"prefix"}
or \code{"suffix"}. By default, \code{" "} will be used.}

\item{name}{A string define the PATH environment variable name. You
can use this to define other \code{PATH}-like environment variable such as
\code{PYTHONPATH}.}
}
\value{
\itemize{
\item \code{cmd_wd}: The \code{command} object itself, with working directory updated.
}

\itemize{
\item \code{cmd_envvar}: The \code{command} object itself, with running environment
variable updated.
}

\itemize{
\item \code{cmd_envpath}: The \code{command} object itself, with running environment
variable \code{name} updated.
}

\itemize{
\item \code{cmd_on_start}: The \code{command} object itself, with the start code updated.
}

\itemize{
\item \code{cmd_on_exit}: The \code{command} object itself, with the exit code updated.
}

\itemize{
\item \code{cmd_on_fail}: The \code{command} object itself, with the failure code updated.
}
}
\description{
Setup the context for the command
}
\section{Functions}{
\itemize{
\item \code{cmd_wd()}: define the working directory.

\item \code{cmd_envvar()}: define the environment variables.

\item \code{cmd_envpath()}: define the \code{PATH}-like environment variables.

\item \code{cmd_on_start()}: define the start code of the command

\item \code{cmd_on_exit()}: define the exit code of the command

\item \code{cmd_on_fail()}: define the failure code of the command

}}
\seealso{
\itemize{
\item \code{\link[=cmd_run]{cmd_run()}}/\code{\link[=cmd_help]{cmd_help()}}/\code{\link[=cmd_background]{cmd_background()}}
\item \code{\link[=cmd_parallel]{cmd_parallel()}}
}
}
