% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-conda.R
\name{conda}
\alias{conda}
\title{Run conda}
\usage{
conda(subcmd = NULL, ..., conda = NULL)
}
\arguments{
\item{subcmd}{Sub-Command of conda.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{conda} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \code{cmd_help(conda())}.}

\item{conda}{A string of path to \code{conda} command.}
}
\value{
A \code{command} object.
}
\description{
Run conda
}
\seealso{
\itemize{
\item \code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}/\code{\link[=cmd_on_exit]{cmd_on_exit()}}
\item \code{\link[=cmd_run]{cmd_run()}}/\code{\link[=cmd_help]{cmd_help()}}/\code{\link[=cmd_background]{cmd_background()}}
}

Other \code{commands}:
\code{\link{allele_counter}()},
\code{\link{cellranger}()},
\code{\link{fastq_pair}()},
\code{\link{gistic2}()},
\code{\link{kraken2}()},
\code{\link{kraken_tools}()},
\code{\link{perl}()},
\code{\link{pyscenic}()},
\code{\link{python}()},
\code{\link{samtools}()},
\code{\link{seqkit}()},
\code{\link{trust4}()}
}
\concept{command}
