% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-.R
\name{exec}
\alias{exec}
\title{Invoke a System Command}
\usage{
exec(cmd, ...)
}
\arguments{
\item{cmd}{Command to be invoked, as a character string.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \code{cmd} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link{shQuote}}.}
}
\value{
A \code{command} object.
}
\description{
Invoke a System Command
}
\examples{
cmd_run(exec("echo", "$PATH"))
}
\seealso{
\itemize{
\item \code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}
\item \code{\link[=cmd_run]{cmd_run()}}/\code{\link[=cmd_background]{cmd_background()}}/\code{\link[=cmd_help]{cmd_help()}}
}
}
