% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-pyscenic.R
\name{pyscenic}
\alias{pyscenic}
\title{Run pyscenic}
\usage{
pyscenic(subcmd = NULL, ..., pyscenic = NULL)
}
\arguments{
\item{subcmd}{Sub-Command of pyscenic.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \verb{pyscenic subcmd} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link{shQuote}}. Details see: \verb{cmd_help(pyscenic subcmd())}.}

\item{pyscenic}{A string of path to \code{pyscenic} command.}
}
\value{
A \code{command} object.
}
\description{
Run pyscenic
}
\references{
\url{https://github.com/aertslab/pySCENIC}
}
