% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-pyscenic.R
\name{pyscenic}
\alias{pyscenic}
\title{Run pyscenic}
\usage{
pyscenic(subcmd = NULL, ..., pyscenic = NULL)
}
\arguments{
\item{subcmd}{Sub-Command of pyscenic.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \verb{pyscenic subcmd} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \verb{cmd_help(pyscenic subcmd())}.}

\item{pyscenic}{A string of path to \code{pyscenic} command.}
}
\value{
A \code{command} object.
}
\description{
Run pyscenic
}
\seealso{
\itemize{
\item \url{https://github.com/aertslab/pySCENIC}
\item \code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}
\item \code{\link[=cmd_run]{cmd_run()}}/\code{\link[=cmd_help]{cmd_help()}}/\code{\link[=cmd_background]{cmd_background()}}
}

Other \code{commands}:
\code{\link{allele_counter}()},
\code{\link{cellranger}()},
\code{\link{fastq_pair}()},
\code{\link{gistic2}()},
\code{\link{kraken2}()},
\code{\link{kraken_tools}()},
\code{\link{perl}()},
\code{\link{python}()},
\code{\link{samtools}()},
\code{\link{seqkit}()},
\code{\link{trust4}()}
}
\concept{command}
