% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd-seqkit.R
\name{seqkit}
\alias{seqkit}
\title{Run seqkit}
\usage{
seqkit(subcmd = NULL, ..., seqkit = NULL)
}
\arguments{
\item{subcmd}{Sub-Command of seqkit.}

\item{...}{<\link[rlang:dyn-dots]{dynamic dots}> Additional arguments passed to \verb{seqkit subcmd} command. Empty arguments are automatically trimmed. If a single argument, such as a file path, contains spaces, it must be quoted, for example using \code{\link[=shQuote]{shQuote()}}. Details see: \verb{cmd_help(seqkit subcmd())}.}

\item{seqkit}{A string of path to \code{seqkit} command.}
}
\value{
A \code{command} object.
}
\description{
Run seqkit
}
\seealso{
\itemize{
\item \url{https://bioinf.shenwei.me/seqkit/}
\item \code{\link[=cmd_wd]{cmd_wd()}}/\code{\link[=cmd_envvar]{cmd_envvar()}}/\code{\link[=cmd_envpath]{cmd_envpath()}}
\item \code{\link[=cmd_run]{cmd_run()}}/\code{\link[=cmd_help]{cmd_help()}}/\code{\link[=cmd_background]{cmd_background()}}
}

Other \code{commands}:
\code{\link{allele_counter}()},
\code{\link{cellranger}()},
\code{\link{conda}()},
\code{\link{fastq_pair}()},
\code{\link{gistic2}()},
\code{\link{kraken2}()},
\code{\link{kraken_tools}()},
\code{\link{perl}()},
\code{\link{pyscenic}()},
\code{\link{python}()},
\code{\link{samtools}()},
\code{\link{trust4}()}
}
\concept{command}
